LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/transformations/2d.transformation.translate.nonfinite-expected.txt

2d.transformation.translate.nonfinite
translate() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS translate() with Infinity/NaN is ignored 
